Phase 59: RISC-V and x86 workstation profiles with PCI IRQ fixes
Bring full workstation smoke (including hostfwd HTTP curl) to qemu_virt_riscv64 and x86_64_generic by adding arch-specific profiles/templates and fixing x86 combo virtio-pci IRQ ack/RX drain so FS and net no longer starve each other.

+++ userspace/pds/log-server/src/main.rs
@@ -84,7 +84,7 @@ impl HandlerImpl {
         let (use_tag, use_tag_len) = if tag_len > 0 {
             (&tag[..tag_len as usize], tag_len)
         } else {
-            (fallback, fallback_len as u8)
+            (fallback, fallback_len)
         };
 
         let mut line = LogLine {
